Moved OplogMonitorThread to a seperate file instead of being an inner class of BlackboardClient. As a result OplogMonitorThread now keeps a copy of the subscriptions and needs to be passed the used Mongo client.
